backend/pyproject.toml
[tool.poetry]
name = "app.shubraj.com-backend"
version = "1.0.0"
description = "backend for app.shubraj.com"
authors = ["Shub Raj Lama <hello@shubraj.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
dnspython = "^2.7.0"
pillow = "^11.0.0"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
